Implement telemetry in pkgsign, so we can gauge usage and adoption
Telemetry in pkgsign records when it's used to verify or sign packages. Data such as package names and versions are only included if the package is not private, as per package.json.

Telemetry can be completely disabled by setting the environment variable `DISABLE_PKGSIGN_TELEMETRY=true`.

This also bumps the version of the dev branch to `0.2.0-alpha.1`.